+++ src/graph_builder/backend/fallback/allocator.py
@@ -4,8 +4,8 @@ import numpy as np
 
 from graph_builder.graph.operator import Operator
 from graph_builder.graph.variable import Variable
-from graph_builder.graph.variables.attributes.constant import Constant as ConstantAttribute
-from graph_builder.graph.variables.constant import Constant
+from graph_builder.graph.variables.attributes.constant import Constant
+from graph_builder.graph.variables.constant_variable import ConstantVariable
 from graph_builder.optimizer import util
 from graph_builder.util import json
 
@@ -74,7 +74,7 @@ class Allocator:
         variables = util.listup_variables(graph, remove_alias=True)
         for i, v in enumerate(variables):
             v.parameters["name"] = f"v{i}"
-        constants = set(util.filter_nodes(variables, ConstantAttribute))  # type: Set[Constant]
+        constants = set(util.filter_nodes(variables, Constant))  # type: Set[ConstantVariable]
         variables = variables.difference(constants)
 
         variables = list(variables)
@@ -85,7 +85,7 @@ class Allocator:
         return variables_layout, constants_layout, data
 
     @classmethod
-    def allocate_constants(cls, constants: List[Constant]) -> Tuple[MemoryLayout, np.ndarray]:
+    def allocate_constants(cls, constants: List[ConstantVariable]) -> Tuple[MemoryLayout, np.ndarray]:
         layout = MemoryLayout()
 
         for constant in constants:
