+++ src/graph_builder/backend/webgpu/optimize_rules/adjust_conv_weight_data_order.py
@@ -1,7 +1,7 @@
 from graph_builder.graph.operator import Operator
 from graph_builder.graph.operators.convolution2d import Convolution2D
 from graph_builder.graph.variables.attributes.order import OrderHWCN
-from graph_builder.graph.variables.constant import Constant
+from graph_builder.graph.variables.constant_variable import ConstantVariable
 from graph_builder.optimizer import util
 from graph_builder.optimizer.optimize_rule import OptimizeRule
 
@@ -19,7 +19,7 @@ class AdjustConvWeightDataOrder(OptimizeRule):
 
             op: Convolution2D
 
-            w = op.inputs["w"]  # type: Constant
+            w = op.inputs["w"]  # type: ConstantVariable
             if w.axis_order == OrderHWCN:
                 continue
